sdx2_encode.c

+++ src/sdx2_encode.c
@@ -163,8 +163,8 @@ sdx2_encode_stereo(const s16 *ibuf_,
 		
       if(i)
         {
-          comp_sample = encode_sample((s32)curr_left_sample,
-                                      (s32)prev_left_sample);
+          comp_sample = encode_sample(curr_left_sample,
+                                      prev_left_sample);
         }
       else 
         {
